Add FakeNavigator and consolidate coe fake modules
