+++ core-testing/src/main/java/com/jraska/github/client/FakeCoreAndroidModule.kt
@@ -18,30 +18,6 @@ object FakeCoreAndroidModule {
   @Provides
   @Singleton
   fun provideNavigator(): Navigator {
-    return object : Navigator {
-      override fun launchOnWeb(httpUrl: HttpUrl) {
-        TODO("Not yet implemented")
-      }
-
-      override fun startUserDetail(login: String) {
-        TODO("Not yet implemented")
-      }
-
-      override fun startRepoDetail(fullPath: String) {
-        TODO("Not yet implemented")
-      }
-
-      override fun showSettings() {
-        TODO("Not yet implemented")
-      }
-
-      override fun showAbout() {
-        TODO("Not yet implemented")
-      }
-
-      override fun startConsole() {
-        TODO("Not yet implemented")
-      }
-    }
+    return Fakes.recordingNavigator()
   }
 }
