+++ core-testing/src/main/java/com/jraska/github/client/Fakes.kt
@@ -29,4 +29,8 @@ object Fakes {
   fun recordingAnalyticsProperty(): RecordingAnalyticsProperty {
     return RecordingAnalyticsProperty()
   }
+
+  fun recordingNavigator() : RecordingNavigator {
+    return RecordingNavigator()
+  }
 }
