+++ feature/users/src/test/java/com/jraska/github/client/users/di/TestUsersComponent.kt
@@ -1,10 +1,6 @@
 package com.jraska.github.client.users.di
 
-import com.jraska.github.client.Config
-import com.jraska.github.client.FakeConfigModule
-import com.jraska.github.client.FakeCoreAndroidModule
-import com.jraska.github.client.FakeCoreModule
-import com.jraska.github.client.http.FakeHttpModule
+import com.jraska.github.client.FakeModules
 import com.jraska.github.client.users.UsersModule
 import com.jraska.github.client.users.UsersViewModel
 import dagger.Component
@@ -13,8 +9,7 @@ import javax.inject.Singleton
 
 @Singleton
 @Component(
-  modules = [FakeCoreModule::class, FakeConfigModule::class,
-    UsersModule::class, FakeCoreAndroidModule::class, FakeHttpModule::class]
+  modules = [FakeModules::class, UsersModule::class]
 )
 internal interface TestUsersComponent {
   fun usersViewModel(): UsersViewModel
